error handing for feign

+++ cashier/src/main/java/org/openexchange/service/CashierService.java
@@ -3,5 +3,5 @@ package org.openexchange.service;
 import java.math.BigDecimal;
 
 public interface CashierService {
-    BigDecimal exchange(String source, String target, BigDecimal amount);
+    BigDecimal exchange(String source, String target, double amount);
 }
